chore: move update category serializer file
