+++ src/controller/operators/updateCategoryOperator.ts
@@ -3,7 +3,7 @@ import { injectable, inject } from 'inversify'
 import { left, right } from '../../framework/shared/either'
 import { AbstractOperator } from './abstractOperator'
 import { UpdateCategoryUseCase } from '../../business/useCases/updateCategoryUseCase'
-import { InputUpdateCategory, OutputUpdateCategory } from '../serializers/inputUpdateCategory'
+import { InputUpdateCategory, OutputUpdateCategory } from '../serializers/categories/inputUpdateCategory'
 
 @injectable()
 export class UpdateCategoryOperator extends AbstractOperator<InputUpdateCategory, OutputUpdateCategory> {
