+++ src/controller/serializers/categories/inputUpdateCategory.ts
@@ -1,9 +1,9 @@
 import { IsNotEmpty, IsHexColor, IsOptional, IsString } from 'class-validator'
 
-import { ICategoryEntity } from '../../domain/entities/categoryEntity'
-import { Either } from '../../framework/shared/either'
-import { IError } from '../../framework/shared/iError'
-import { Validatable } from './abstractValidatable'
+import { ICategoryEntity } from '../../../domain/entities/categoryEntity'
+import { Either } from '../../../framework/shared/either'
+import { IError } from '../../../framework/shared/iError'
+import { Validatable } from '../abstractValidatable'
 
 export class InputUpdateCategory extends Validatable<InputUpdateCategory> {
     @IsNotEmpty()
