+++ src/framework/functions/categories/updateCategory.ts
@@ -6,7 +6,7 @@ import { httpHandler } from '../../utility/httpHandler'
 import { container } from '../../shared/ioc/container'
 import { httpResponse } from '../../utility/httpResponse'
 import { UpdateCategoryOperator } from '../../../controller/operators/updateCategoryOperator'
-import { InputUpdateCategory } from '../../../controller/serializers/inputUpdateCategory'
+import { InputUpdateCategory } from '../../../controller/serializers/categories/inputUpdateCategory'
 
 export const handler = httpHandler(async (event: APIGatewayProxyEvent, context: Context) => {
     context.callbackWaitsForEmptyEventLoop = false
